feat: improve Discord username error correction
